+++ backend/authentication/views/updateUserInfo.py
@@ -71,134 +71,6 @@ def get_profile_info(request):
   else:
     return Response(status=405)
 
-@api_view(['GET'])
-@authentication_classes([TokenAuthentication])
-@permission_classes([IsAuthenticated])
-def get_stats(request, userId):
-  try:
-    user = get_object_or_404(User, id=userId)
-    gameWon = UserStat.objects.filter(player=user, isWinner=True)
-    nbrGameWon = gameWon.count()
-    allGames = UserStat.objects.filter(player=user).order_by('date')
-    nbrGames = allGames.count()
-
-    mapPercentages = {}
-    modePercentages = {}
-
-    # Win rate
-    percentageGameWon = round((nbrGameWon * 100) / user.nbr_match, 1) if user.nbr_match > 0 else 0
-    percentageGameLost = round(100 - percentageGameWon, 1)
-
-    # Dashes / PoweredUsed
-    sums = allGames.aggregate(totalDashes=Sum('nbDashes'), totalPoweredUsed=Sum('nbPoweredUsed'))
-
-    totalDashes = sums['totalDashes'] or 0
-    totalPoweredUsed = sums['totalPoweredUsed'] or 0
-
-    total = totalDashes + totalPoweredUsed
-    if total > 0:
-      dashesPercentage = (totalDashes / total) * 100
-      poweredUsedPercentage = (totalPoweredUsed / total) * 100
-    else:
-      dashesPercentage = 0
-      poweredUsedPercentage = 0
-
-    # Efficiency rate
-    allGames_annotated = allGames.annotate(
-      efficiency_ratio=ExpressionWrapper(
-        F('pointsScored') / 
-        Case(
-            When(pointsTaken__gt=0, then=F('pointsTaken')),
-            default=1.0,
-            output_field=FloatField()
-        ),
-        output_field=FloatField()
-      )
-    )
-
-    efficiencyRatios = list(allGames_annotated.values_list('efficiency_ratio', flat=True))
-
-    # Current streak
-    maxStreak = 0
-    currentStreak = 0
-
-    for game in allGames:
-      if game.isWinner:
-        currentStreak += 1
-        if currentStreak > maxStreak:
-          maxStreak = currentStreak
-      else:
-          currentStreak = 0
-    
-    spaceMap = 0
-    dragonMap = 0
-    skyMap = 0
-    oceanMap = 0
-    classicMode = 0
-    spinOnlyMode = 0
-    powerlessMode = 0
-
-    for game in allGames:
-      
-      # Map counter
-      print("map game", game.mapGame)
-      if game.mapGame == "oceanMap":
-        oceanMap += 1 
-      elif game.mapGame == "spaceMap":
-          spaceMap += 1
-      elif game.mapGame == "dragonMap":
-          dragonMap += 1
-      elif game.mapGame == "skyMap":
-          skyMap += 1
-      
-      # Mode counter
-      print("mode game", game.modeGame)
-      if game.modeGame == "CLASSIC":
-        classicMode += 1
-      elif game.modeGame == "SPINONLY":
-        spinOnlyMode += 1
-      elif game.modeGame == "POWERLESS":
-        powerlessMode += 1
-    
-    # Map percentage
-    oceanMapPercentage = round((oceanMap / nbrGames) * 100, 1) if user.nbr_match > 0 else 0
-    spaceMapPercentage = round((spaceMap / nbrGames) * 100, 1) if user.nbr_match > 0 else 0
-    dragonMapPercentage = round((dragonMap / nbrGames) * 100, 1) if user.nbr_match > 0 else 0
-    skyMapPercentage = round((skyMap / nbrGames) * 100, 1) if user.nbr_match > 0 else 0
-
-    # Mode percantage
-    classicModePercentage = round((classicMode / nbrGames) * 100, 1) if user.nbr_match > 0 else 0
-    spinOnlyModePercentage = round((spinOnlyMode / nbrGames) * 100, 1) if user.nbr_match > 0 else 0
-    powerlessModePercentage = round((powerlessMode / nbrGames) * 100, 1) if user.nbr_match > 0 else 0
-
-    mapPercentages = {
-      "oceanMap": oceanMapPercentage,
-      "spaceMap": spaceMapPercentage,
-      "dragonMap": dragonMapPercentage,
-      "skyMap": skyMapPercentage
-    }
-
-    modePercentages = {
-      "classicMode": classicModePercentage,
-      "spinOnlyMode": spinOnlyModePercentage,
-      "powerlessMode": powerlessModePercentage
-    }
-
-    return Response({
-      'percentageGameWon': percentageGameWon, 
-      'percentageGameLost': percentageGameLost,
-      'dashesPercentage': dashesPercentage,
-      'poweredUsedPercentage': poweredUsedPercentage,
-      'efficiencyRatios': efficiencyRatios,
-      'nbrGames': nbrGames,
-      'currentStreak': currentStreak,
-      'mapPercentages': mapPercentages,
-      'modePercentages' : modePercentages
-    })
-
-  except User.DoesNotExist:
-    return Response({'status': "Not found", 'error': "L'utilisateur avec cet identifiant n'existe pas."}, status=404)
-
 @api_view(['POST'])
 @authentication_classes([TokenAuthentication])
 @permission_classes([IsAuthenticated])
