+++ backend/html/index.html
@@ -252,6 +252,7 @@
       <div class="validateDeleteText">Enter your password</div>
       <div class="changeInfoBar" style="height: 20%;">
         <input id="enterPasswordInput" name="password" class="text-input3" style="font-size: 20px;" spellcheck="false" type="password" placeholder="Password"></input>
+        <div id="errorLogGuest"></div>
       </div>
       <div class="buttonContainer" style="height: 10%;">
         <div id="arenaBackLogInButton" class="validateDeleteButton" style="border: 4px solid #ff3737; background-color: #ff373771; color: #ff3737;">back</div>
