+++ backend/static/html/js/stats.js
@@ -90,9 +90,10 @@ export function getUserStats(userId) {
 		// console.log("poweredUsedPercentage", data.poweredUsedPercentage);
 		// console.log("efficencyRatios", data.efficiencyRatios);
 		// console.log("nbrGames", data.nbrGames);
-		console.log("currentStreak", data.currentStreak);
-		console.log("mapPercentages", data.mapPercentages);
-		console.log("modePercentages", data.modePercentages);
+		// console.log("currentStreak", data.currentStreak);
+		// console.log("mapPercentages", data.mapPercentages);
+		// console.log("modePercentages", data.modePercentages);
+		console.log("efficiency", data.efficiency);
 		drawChart();
 	})
 	.catch(error => {
